[ci] Consolidate Verible linting workflow into one stage
Running the verible linter and adding review comments to the pull request
previously had to be done in two stages:

1. Triggered on the pull request - prepare config and waiver files as artifacts.
2. Running on the repo's HEAD - run Verible and add review comments.

This was required because Actions running in the context of the pull request
did not have write permissions to add comments to pull requests.

This is now possible with the `pull_request_target` event, which triggers
when pull requests change, but runs in the context of the repo's HEAD and
has the permissions to create comments.

See https://github.com/lowRISC/ibex/pull/1427 and
https://github.com/chipsalliance/verible-linter-action/pull/31 for details.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>